Separate combinator and pure components
